STM32: ADCv2: set DDS bit only in circular mode
According to DS:
At the end of the last DMA transfer (number of transfers configured in the DMA controller’s
DMA_SxNTR register):
- No new DMA request is issued to the DMA controller if the DDS bit is cleared to 0 in the
ADC_CR2 register (this avoids generating an overrun error). However the DMA bit is
not cleared by hardware. It must be written to 0, then to 1 to start a new transfer.
- Requests can continue to be generated if the DDS bit is set to 1. This allows
configuring the DMA in double-buffer circular mode.